# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Self-Service Totem",
        "description": "Documentation for Self Service Totem API",
        "version": "1.0.0",
        "contact": {
            "name": "API Support Team",
            "url": "https://github.com/evilfeeh/self-service-totem/issues"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base path"
        }
    ],
    "paths": {
...
        "/order/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "tags": ["Orders"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to find",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/OrderOutputDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "summary": "Update an existing order",
                "tags": ["Orders"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/OrderUpdateDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "updated successfully"
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/order": {
            "get": {
                "summary": "List orders",
                "tags": ["Orders"],
                "responses": {
                    "200": {
                        "description": "Orders found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderOutputDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: Unable to find orders"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "summary": "Create a new order",
                "tags": ["Orders"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "type": "object",
                            "schema": {
                                "$ref": "#/components/schemas/OrderInputDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "Order created successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request: Missing or invalid parameters"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderInputDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "OrderUpdateDTO": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "Recebido",
                            "Em preparação",
                            "Pronto",
                            "Finalizado"
                        ]
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                }
            },
            "OrderOutputDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "customer": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "productId": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "total": {
                        "type": "number"
                    },
                    "closed": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}